+++ packages/webviz-core/package.json
@@ -11,6 +11,7 @@
     "@sentry/browser": "5.11.0",
     "@trust/webcrypto": "0.9.2",
     "antd": "4.1.1",
+    "array-move": "^3.0.1",
     "chart.js": "davidswinegar/Chart.js#0785c29a4c57b3e1b4243f8163df7950a00b2a80",
     "chartjs-plugin-annotation": "0.5.7",
     "chartjs-plugin-datalabels": "davidswinegar/chartjs-plugin-datalabels#b0ade94116d4e273f0fdc17e9528c9a33c796e90",
@@ -61,6 +62,7 @@
     "react-router": "5.0.1",
     "react-router-dom": "5.0.1",
     "react-select": "1.2.1",
+    "react-smooth-dnd": "^0.11.1",
     "react-sortable-hoc": "1.10.1",
     "react-virtualized": "9.21.0",
     "react-virtualized-auto-sizer": "1.0.2",
