+++ packages/webviz-core/src/panels/ScenarioEditor/index.js
@@ -9,17 +9,13 @@
 import _ from "lodash";
 import React, { useState } from "react";
 import { hot } from "react-hot-loader/root";
-import styled from "styled-components";
 import CSVReader from 'react-csv-reader';
-import { makeStyles } from '@material-ui/core/styles';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import Collapse from '@material-ui/core/Collapse';
-import ExpandLess from '@material-ui/icons/ExpandLess';
-import ExpandMore from '@material-ui/icons/ExpandMore';
-import DeleteIcon from '@material-ui/icons/Delete';
 
+import List from '@material-ui/core/List';
+import { Container, Draggable } from 'react-smooth-dnd';
+import arrayMove from 'array-move';
+
+import Scenario from "./Scenario";
 import helpContent from "./index.help.md";
 import Flex from "webviz-core/src/components/Flex";
 import Panel from "webviz-core/src/components/Panel";
@@ -28,15 +24,14 @@ import PanelToolbar from "webviz-core/src/components/PanelToolbar";
 type Config = { errorMessages: Object };
 type Props = { config: Config };
 
-const useStyles = makeStyles((theme) => ({
-  nested: {
-    paddingLeft: theme.spacing(4),
-  },
-}));
 
 function ScenarioEditor({ config }: Props) {
 
   const [scenarios, setScenarios] = useState([]);
+
+  const onDrop = ({ removedIndex, addedIndex }) => {
+    setScenarios(scenarios => arrayMove(scenarios, removedIndex, addedIndex))
+  };
 
   return (
     <Flex col style={{ height: "100%" }}>
@@ -46,11 +41,21 @@ function ScenarioEditor({ config }: Props) {
       </div>
       {
         scenarios.length > 0 ?
-          <ScenarioList scenarios={scenarios} />
+          <List>
+            <Container dragHandleSelector=".drag-handle" lockAxis="y" onDrop={onDrop}>
+            {
+              scenarios.map((scenario, index) =>
+                <Draggable key={index}>
+                  <Scenario key={index} scenario={scenario} />
+                </Draggable>
+              )
+            }
+            </Container>
+          </List>
           :
           <div>
             <CSVReader
-              onFileLoaded={(data) => setScenarios(data)}
+              onFileLoaded={(data) => { setScenarios(data); console.log(data); }}
               parserOptions={{
                 header: true,
                 dynamicTyping: true,
@@ -62,58 +67,6 @@ function ScenarioEditor({ config }: Props) {
     </Flex>
   );
 
-}
-
-function ScenarioList({ scenarios }) {
-  return (
-    <List>
-      {
-        scenarios.map((scenario, index) =>
-          <Scenario key={index} scenario={scenario} />
-        )
-      }
-    </List>
-  )
-}
-
-function Scenario({ scenario }) {
-
-  const classes = useStyles();
-  const [open, setOpen] = useState(false);
-
-  return (
-    <div>
-      <ListItem button onClick={() => setOpen(prevOpen => !prevOpen)} >
-        <ListItemText primary={`${scenario.start_id} - ${scenario.end_id}`} />
-        <DeleteIcon onClick={(e) => { alert(1); e.preventDefault(); }} />
-        {open ? <ExpandLess /> : <ExpandMore />}
-      </ListItem>
-      <Collapse in={open} timeout="auto" unmountOnExit>
-        <List component="div" disablePadding>
-          <ListItem button className={classes.nested}>
-            <ListItemText primary="Starred" />
-          </ListItem>
-          <ListItem button className={classes.nested}>
-            <ListItemText primary="Starred" />
-          </ListItem>
-          <ListItem button className={classes.nested}>
-            <ListItemText primary="Starred" />
-          </ListItem>
-        </List>
-      </Collapse>
-    </div>
-  )
-}
-
-function ScenarioItem({ item }) {
-
-  const classes = useStyles();
-
-  return (
-    <ListItem button className={classes.nested}>
-      <ListItemText primary="Starred" />
-    </ListItem>
-  )
 }
 
 ScenarioEditor.panelType = "ScenarioEditor";
